move datapoints compat to prototype datasets
