+++ torchvision/prototype/datasets/_builtin/cub200.py
@@ -15,8 +15,7 @@ from torchdata.datapipes.iter import (
 )
 from torchdata.datapipes.map import IterToMapConverter
 from torchvision.prototype.datapoints import BoundingBox, Label
-from torchvision.prototype.datapoints._datapoint import Datapoint
-from torchvision.prototype.datasets.utils import Dataset, EncodedImage, GDriveResource, OnlineResource
+from torchvision.prototype.datasets.utils import Dataset, EncodedImage, GDriveResource, GenericDatapoint, OnlineResource
 from torchvision.prototype.datasets.utils._internal import (
     getitem,
     hint_sharding,
@@ -162,7 +161,7 @@ class CUB200(Dataset):
                 format="xyxy",
                 spatial_size=spatial_size,
             ),
-            segmentation=Datapoint(content["seg"]),
+            segmentation=GenericDatapoint(content["seg"]),
         )
 
     def _prepare_sample(
